# docs/swagger.yaml
basePath: /v1
definitions:
  swaggertypes.AIMessageData:
    properties:
      content:
        example: Summarise my tasks for today
        type: string
      created_at:
        type: string
      id:
        example: msg_01jw...
        type: string
      role:
        example: user
        type: string
      session_id:
        example: sess_01jw...
        type: string
    type: object
  swaggertypes.AIMessageListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/swaggertypes.AIMessageData'
        type: array
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.AIMessageResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.AIMessageData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.AISessionData:
    properties:
      created_at:
        type: string
      id:
        example: sess_01jw...
        type: string
      status:
        example: active
        type: string
      title:
        example: New Conversation
        type: string
      updated_at:
        type: string
      user_id:
        example: usr_01jw...
        type: string
    type: object
  swaggertypes.AISessionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/swaggertypes.AISessionData'
        type: array
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.AISessionResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.AISessionData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.EmptyData:
    type: object
  swaggertypes.EmptyResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.EmptyData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.ErrorBody:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: email is required
        type: string
    type: object
  swaggertypes.ErrorResponse:
    properties:
      data:
        type: object
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  swaggertypes.MessageData:
    properties:
      message:
        example: logged out
        type: string
    type: object
  swaggertypes.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.MessageData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  swaggertypes.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  swaggertypes.SendMessageRequest:
    properties:
      content:
        example: Summarise my tasks for today
        type: string
    type: object
...
  swaggertypes.TokensData:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  swaggertypes.TokensResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.TokensData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.UserData:
    properties:
      created_at:
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: usr_01jw...
        type: string
      name:
        example: Jane Doe
        type: string
      timezone:
        example: UTC
        type: string
      updated_at:
        type: string
    type: object
...
  swaggertypes.UserResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.UserData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the Nicoflow productivity platform
  title: Nicoflow API
  version: "1.0"
paths:
  /ai/sessions:
    get:
      description: Returns all AI conversation sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.AISessionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List AI sessions
      tags:
      - ai
    post:
      description: Creates a new AI conversation session for the authenticated user
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggertypes.AISessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an AI session
      tags:
      - ai
  /ai/sessions/{id}:
    delete:
      description: Deletes an AI conversation session owned by the authenticated user
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an AI session
      tags:
      - ai
  /ai/sessions/{sessionId}/messages:
    get:
      description: Returns all messages in the specified AI conversation session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.AIMessageListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List messages in a session
      tags:
      - ai
    post:
      consumes:
      - application/json
      description: Sends a message in an AI conversation session and returns the assistant
        reply
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Message content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggertypes.AIMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send a message
      tags:
      - ai
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access + refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the provided refresh token
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out
      tags:
      - auth
  /auth/logout-all:
    post:
      description: Invalidates all refresh tokens for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out all devices
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access + refresh token
        pair (single-use rotation)
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: Registration credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggertypes.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"